requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
flit==1.3
idna==2.8
importlib-metadata==0.23
lark-parser==0.7.8
lispy-fabio==0.1.0
meme==0.1.0
more-itertools==7.2.0
packaging==19.2
pathlib2==2.3.5
Pillow==6.2.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.4
pytest==5.2.2
pytoml==0.1.21
requests==2.22.0
six==1.13.0
urllib3==1.25.6
wcwidth==0.1.7
zipfile36==0.1.3
zipp==0.6.0
